pyproject.toml
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "price-estimator"
packages = [{ include = "predict.py", from = "src" }]
readme = "README.md"
version = "0.1.0"

[tool.poetry.scripts]
price-predictor = "predict:main"

[tool.poetry.dependencies]
grpcio = "^1.62.1"
hydra-core = "1.3"
lightgbm = "^4.3.0"
minio = "^7.2.5"
numpy = "^1.26.4"
pandas = "^2.2.2"
python = "^3.11"
rootutils = "^1.0.7"
scikit-learn = "^1.4.2"
python-dotenv = "^1.0.1"
protobuf = "^5.26.1"

[tool.poetry.group.protos.dependencies]
grpcio-tools = "^1.62.1"


[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
